feat(quest-detail): finish revamp
