+++ package.json
@@ -29,7 +29,7 @@
   "devDependencies": {
     "@chromatic-com/storybook": "^1.9.0",
     "@hyperplay/chains": "^0.3.0",
-    "@hyperplay/ui": "file:/Users/eliobriceno/Documents/hyperplay/ui-library/hyperplay-ui-1.8.17-alpha.0.tgz",
+    "@hyperplay/ui": "^1.8.17",
     "@hyperplay/utils": "^0.2.7",
     "@mantine/core": "^7.5.1",
     "@storybook/addon-essentials": "^8.3.4",
