+++ src/components/RewardWrapper/index.module.scss
@@ -1,0 +1,9 @@
+.rewardContainer {
+  display: flex;
+  gap: var(--space-md);
+  align-items: flex-start;
+}
+
+.alertCard {
+  flex: 1;
+}
