+++ src/types/quests.ts
@@ -1,0 +1,4 @@
+import { QuestReward } from '@hyperplay/ui'
+import { Reward } from '@hyperplay/utils'
+
+export type UseGetRewardsData = QuestReward & Reward
